do not reset actions that have started to run
When an action starts to run, it gets the 'in-progress' status in the DB. These actions
will not be reset to 'pending' when they timeout. After a (default) 30 minute timeout,
these actions will be set to the new 'failed' status. Plugins can decide from there what
to do with the actions.

Fixes #15, Updates #16