Update get_datetime_object()
To accept a timezone parameter, which defaults to UTC, so that
it can be used throughout the codebase to ensure that all timezones
are in UTC by default, rather than the PHP timezone (which can change
on different WordPress installs if other plugins are calling the
date_default_timezone_set() method).

It also always use new DateTime() instead of using a default of
time() for consistency.