Base get_date() on gmt values
Introduce net get_date_gmt() and use this as the base for get_date()
so that if the WordPress or PHP timezone changes, the stored values
on the scheduled action post will not be interpretted incorrectly
(i.e. in the current timezone instead of the timezone active when they
were set).