Introduce as_get_datetime_object()
Instead of using ActionScheduler::get_datetime_object() because it's
more concise and should be available in the procedural API instead
of being within the ActionScheduler, which is intended to setup
the scheduler, not provide utility functions (as clearly indicated
by the lack of unit tests for it).

Also deprecates ActionScheduler::get_datetime_object().

This function uses the `as` prefix becuase it's more unique to Action
Scheduler (and probably should have been chosen initially) so reduces
the risk of conflicts if WC introduces it's own date/time API.