make option textareas resizable
This has been a frequent request from customers. It is indeed very
annoying to be stuck with a fixed width textarea when you need to enter
long URLs and screen estate actually does allow for resizing. It may be
a WordPress default to only be able to resize vertically, but
WordPress’ trextareas are wider by default. Giving people control over
textarea width seems reasonable here.