license validation: removed obsolete tabs
When the license has not been validated yet, the support form does not
send any messages, so the Support tab really should not appear.
Removed the FAQ tab as well, because it does not provide any value to a
person who tries to activate their license.
The license validation process should be revisited in 3.0.