Rewrite human_interval()
To avoid duplication and use human friendly variable names. Also to
make it usable for displaying the interval for a scheduled date in
a more descriptive way than human_time_diff(), because we can include
two periods, like "3 months 4 days", instead of the "3 months" that
human_time_diff() would return.