Set and use a start time when calculating the request execution time
If there are other tasks in WP Cron that run prior to AS, then AS will have much less than the maximum available request time,
and in some cases, much less (e.g. if another plugin is running 30 seconds worth of tasks, AS will only have 30 seconds, not 60).

For this reason set the start time as early in the WP init as possible and use it when calculating the execution time.