Fix JS Errors When Using HubSpot Forms
Fixes a JS and display error when using Hubspot Forms while the inlined JS can be combined the file its self-needs to be excluded (mostly because it includes the required data for the inline form to be loaded. 

Example of full code. 

<script type="text/javascript" src="https://js.hsforms.net/forms/shell.js"></script>
<script>
  hbspt.forms.create({
	portalId: "your-id",
	formId: "your-form-id"
});
</script>

No you can't exclude the inline JS and include the external file because the shell.js declares the hbspt.forms class.