Excludes Dynamic NewsPaper Theme JS
This excludes the inline JS from NewsPaper for the view counter. This JS is dynamic in the sense that it is connected to a unique post ID 
                jQuery().ready(function jQuery_ready() {
                    tdAjaxCount.tdGetViewsCountsAjax("post","[1101]");
                });
            
Because of this, it spins up a bunch of files, for every single post (even though the only part changing is the post ID for each post.