Deprecate get_local_timezone() APIs
Because we can't be sure we'll always be able to get a Timezone String,
meaning we can't be sure we'll always be able to instantiate a valid
timezone for the timezone the site is operating in. Specifically, if a
UTC offset is set in WordPress, but no timezone string and that UTC
offset has no matching timezone string, then we want to be free to use
the UTC offset on the DateTime.