Remove test_local_timezone_offsets_utc()
GMT/UTC offsets do not map 1-to-1 to timezone strings, because more
than one place may fall in the same timezone, e.g. in UTC-12, both
Pacific/Kwajalein and Pacific/Fiji are valid timezone strings.

The timezone string mapping is also no longer needed now that we
fallback to use UTC offset instead of trying to map an offset to
a timezone string, so we don't need to test for it either.