Convert DateTime to ActionScheduler_DateTime
To make it easier for existing methods to take advantage of set_local_timezone()
while maintaining backward compatibility when they accept a DateTime param.

For example, the create_wp_comment() method or get_scheduled_date_string_local()
method both accept a DateTime as a param, and then call set_local_timezone() on
that param.

This could be handled individual in those methods, but it avoids duplicate code
accepting a DateTime and convert it just before it's required.