Excludes ARForms Plugin Inline JS (#1231)
Fixes an issue with the ARForms plugin highlighted in issue #1188 the forms uses inline JS as opposed to HTML to call the form. This just excludes the JS as was highlighted. For reference, this is the plugin in question.

https://codecanyon.net/item/arforms-wordpress-form-builder-plugin/6023165
(cherry picked from commit ecb4cb0ecbefe0a720187c1be6df670ee151f7a5)