Deprecate max execution time
And instead, use a consistent (but filterable) 30 seconds as both
the maximum execution time, and the runner time.

Reducing the time limit to 30 seconds, from an unknown amount in many
cases, helps reduce the likelyhood of timeouts affecting batches.

This is particularly important now that Action Scheduler will loop over
batches in the available time and memory limits (with the merge of #192)
instead of simply attempting one batch and then completing, meaning any
time there are more than the default batch size of actions sitting in
the queue, actions will be processed until we run out of actions, time
or memory..

Unifying the time limit API also fixes a few issues:

1. set_time_limit() was being called with a value that was different
   to the return value of get_maximum_execution_time() in many cases.
   This was at best ineffecient, and at worst dangerous as we could
   perceive of having two different time limits.
2. As one consequence of the above, ini_get( 'max_execution_time' ) -
   the default value when not on a known host - would presumably
   return 600, but that was intended to more of an optimistic increase
   not a reliably value for the amount of time we could consume.
3. To consistently change the amount of time allowed for a queue, the
   a callback on the 'action_scheduler_queue_runner_time_limit' filter
   was no longer enough (it would have been prior to 2.0.0), instead
   both it and 'action_scheduler_maximum_execution_time' needed to have
   a callback.

Fixes #207.