Increase likelihood of timeout
Having reviewed data on action processing across various sites, it's
clear estimating with a factor of 2 is not sufficient - there is too
much variance.

Increasing this to a factor of 3 increases the likelihood of time
being exceeded, so will reduce the rate of batch processing, but in
return, there will be a much lower risk of a batch being halted by
a timeout.