Display hook on WP CLI output
Because the action has been instantiated by the time the
'action_scheduler_after_execute' hook is triggered, it's
possible to pass this instance of the action to callbacks.

This makes it possible to display the hook name (or other
action properties) on that callback.

It's not possible to pass the action instance on other hooks
used for WP CLI output, like 'action_scheduler_before_execute'
or 'action_scheduler_failed_execution', because there is no
guarantee at that stage that the action will have been
successfully instantiated in the case of the later, and moving
the instantiation to happen prior to the former would be a
breaking change. Because of this, its only be passed to the
'action_scheduler_after_execute' hook.

Fixes #225