Add ActionScheduler_Store::has_pending_actions_due
As a helper method for checking if there are pending actions that
need to be run. For now, it's just a convenience wrapper for the
query_actions() call. In future, it may be handy to cache the
result (though this isn't done yet as caching that info has the
potential to introduce dangerous bugs).

Added to based ActionScheduler_Store to make the method backward
compatible.