Add locking APIs
To replace transient lock system in ActionScheduler_QueueRunner.

This makes it possible to get more information about the lock from
other code, like ActionScheduler_ListTable so that it can display
a notice about the async queue processing when no lock is set or
when the lock is set but will expire soon.

It also makes it possible to swap out the lock for a different
system if required, rather than having a hardcoded transient used
as the data store in ActionScheduler_QueueRunner.

To live up to the "lock" name more than transients would support
(which provide an 'up-to' expiration, not a guaranteed expiration)
this patch also swaps to use autoloaded options in place of the
transient. This also made sense given the value of the option was
set to the expiration date so that it could be used to display
the amount of time before the next queue would begin.