Fix concurrent queue notice
The notice displayed to alert an admin about the maximum number of
queues being processed used the term "batches", which is incorrect.
Even when the queue concurrency has been reached, new batches may
be claimed.

It said "No actions will be processed until", which is incorrect -
actions continue to be processed by the active queues, but no
additional queues will be initiated.