Improve fetch exception handling
The previous code caught exceptions and then sort of "nulled" the
action by setting default values for args and schedule. However,
the returns action would still be a valid action because of the
call to get_stored_action(), which never returns an instance of
ActionScheduler_NullAction. This can lead to bugs later down the
line.

When data is corrupted, fetch_action() should consistently return
an instance of ActionScheduler_NullAction, like it already is when
passed an invalid ID.