Fix recurring schedules, e.g. Cron and Interval schedules
Recurring schedules issues, as outlined in #330.

This patch fixes recurring schedules, and in the process, refactors
all schedules to avoid duplicate code and ambiguity of properties
and parameters.

This patch introduces to important new base classes:
 * ActionScheduler_Abstract_Schedule and
 * ActionScheduler_Abstract_RecurringSchedule

These are then used to implement all other schedule types - cron,
recurring/interval and simple. This helps unify the data on each
schedule. For example, schedules used to have different property
names referring to equivalent data.

For example, ActionScheduler_IntervalSchedule::start_timestamp
was the same as ActionScheduler_SimpleSchedule::timestamp.

Both of these timestamps actually referring to the date the action
was scheduled, but due to ambiguity in naming, like 'start' or 'date'
or 'first', the data was sometimes used to refer to the scheduled date,
and sometimes the first instance of a recurring schedule. While
unifying the properties, their nomenclature has also been changed
to avoid ambiguity.

Fixes #330 and associated issues - #256, #258, #291, #293 and #296.