Add ActionScheduler_CanceledSchedule
Now that the scheduled date and future / next date calculations
have been separated, we can create a new schedule which will
still return a cancelled action's scheduled date, but never
show a cancelled action as having a next date.

Previously, cancelled actions used the Null schedule, which would
always return NULL for both the next scheduled date, and the
scheduled date of the action, meaning we could never find out
when a cancelled action was originally scheduled to run.