Do not mark action complete in Migration\Scheduler
Because when the migration is run via the Scheduler, the action
will be marked as complete by the QueueRunner and when the migration
is run via Migration_Command::migrate(), the scheduled action should
not be marked 'complete', because the action is not actually run (or
completed). It should instead simply be unscheduled so that no
migration is run in future via the Migration Scheduler once the
WP CLI command has completed the migration.

This minor bug was introduced with a few other changes in SHA: d74bf96a