Maintain property names for serialized schedules
Prior to #333, recurring schedules used different property names
to refer to equivalent data.

For example, IntervalSchedule::start_timestamp was the same as
SimpleSchedule::timestamp.

PR #333 aligned properties and property names for better inheritance.

To guard against the possibility of infinite loops if downgrading to
Action Scheduler < 3.0.0, we need to also store the data with the old
property names so if it's unserialized in AS < 3.0, the schedule
doesn't end up with a 0 recurrence.

Fixes #340.