Loaded Patchwork during the bootstrap to ensure it's ready.
Why?

Patchwork lets us redefine user-defined functions for "mocking" in tests.
We use it with Brain Monkey whenever we mock functions using `when`, `expect`, etc.

Patchwork is fussy. It must be loaded into memory first, before anything else.
If we don't, we can weird some weird behaviors.

This commit loads it as part of the test bootstrapping process.