Added AdminOnly integration test.
Why? To test the admin features, functionality, and screen, we need to be in the Admin before Rocket loads.

Why? is_admin() needs to be set to load up and register all Admin files, Service Providers, etc.

Defining the constant WP_ADMIN and setting it to true during the bootstrap is necessary to ensure the AdminOnly tests can run.