Unmark migration when a plugin is deactivated to prevent action loss.
If a plugin that loads AS 3.x is deactivated, flag for migration completed still remains. This means that if remaining plugins have highest active AS version of 2.x, then in future if a plugin with AS 3.x is installed, migration routine will not kick in because migration flag was already set.

An example of this is when WC Services with loads AS 3.x is disabled and WC 3.9 is enabled on the site, then this effectively downgrades AS from 3.x to 2.x. Then if WC is updated from 3.x to 4.x, action migration will not kick in because migration was marked completed by WC services back when it was updated/installed.

Also, since we cannot reliably determine what AS version will remain after deactivation, hook is triggered for all plugin deactivations.