add flag to force searching the primary store first
AS only uses the hybrid store fetch_action in the queue runners and the list table.
The hybrid store claims actions for the queue runners from the primary store.
The flag ensures that the queue runner receives actions from the primary store
in the event of a duplicate action ID.
If there is a duplicate action ID between stores, the secondary store one will be
inaccessible in the list table until it is migrated.