Ensures regex iterator targets the cache/min/ folder.
It's possible that the extension is part of the domain name or absolute filesystem path.

For example, the hosting username or the domain name could be css.example.org.

By prefixing the regex with the absolute path to the cache/min/, we make sure it matches the items in the right folder.