Changes the design strategy to get the cache/wp-rocket/ root directories.
Why? Faster.

How so?

Before with each URL, we crawled the filesystem. As our query had a wildcard,
i.e. example.org*, we need a tool like RegexIterator or glob() to grab
the root domain + any user caches directories.
Now, that only runs once and is held statically in memory.

Regex/wildcard crawls are expensive. With this design, we only run it when it's needed.