Makes rocket_generate_advanced_cache_file() testable by allowing AdvancedCache to be injected in it.
Think of this dependency injection as a reference to the AdvancedCache object.

This function will be part of AdvancedCache handler. Then it'll have an internal reference via $this. Until then, to make the code testable, we need a way to inject the object in order to set the template to vfs.