Apply font-display: swap in CSS files - Combine ( 2/5 )  (#2841)
Sub-task for epic #2791 
Closes https://github.com/wp-media/wp-rocket/issues/2797

## Part2 - Combine.php
https://github.com/wp-media/wp-rocket/blob/dddb3267bbbce64b6bd9602ab71edeb405f3f1d2/inc/Engine/Optimization/Minify/CSS/Combine.php#L178
Call `WP_Rocket\Engine\Optimization\CSSTrait ::apply_font_display_swap( $css_file_content  )` from combine() function

Unit & Integration tests do not check the content of the minified CSS files, only that the swap of the CSS files with the minified new files happens. So there is no need for altering the unit / integration tests. For apply font-display-swap change the tests from the CSSTrait should be responsible for testing that this code works.

This PR is made from this PR https://github.com/wp-media/wp-rocket/pull/2837 (it needed the CSSTrait) and will be merged in `feature/auto-apply-font-display-swap` branch. Until the PR 2837 is done this PR will contain the changes from the other branch.