Set the PHP version to 7.1 in composer.json for PHP 8 compatibility
When the PHP version in the development system is 8.0 the
"composer install" command will fail saying that some packages
require PHP ^7.0 or ^7.1 but version 8.0 doesn't match.

To fix, this commit adds a "platform" setting in composer.json telling
composer to assume version 7.1 instead of the actual version
in the development system.