Exclude reload_attached_coupons from inline merge
WooCommerce Smart Coupons adds the following inline JS and it changes its URI Component every product. For larger stores where this is the only difference it can cause the cache size to balloon out of control.

Diff checker: https://www.diffchecker.com/m7oUR6P0

I noticed this on a store that had 4k products which for smaller stores this  wouldn't be an issue but it ended up  consuming a very large amount of storage. Excluding the inline JS appears to have fixed the issue.