Fix 4089 Stop applying defer JS exclusions to combine js when defer JS is inactive (#4636)
* Add check for can_defer_js() inside get_excluded()

* Add unit tests for DeferJS::get_exclusions()

After some discussion with Remy on other tests for similar "hard-coded" default exclusion lists, I'm taking a different approach to these tests, so some remarks here. (I'll copy these into the PR as well for further discussion.)

On a recent PR from one of the lvl 2s, I noticed that changes were made to an exclusion list, and tests continued to pass without any modifications. It could pass because, as an exclusion, it didn't change either the expected or the actual HTML without hard-coding the new exclusion into the text fixture as well.

With these kinds of tests, we are not really testing anything other than "I hard-coded what I hard-coded." This is not useful, and has resulted in massively long test fixtures that provide little or no in-code documentation of what is supposed to be under test or why.

It's similar to testing a calculator function by trying to create a fixture that will hard-code every possible combination of integers that might be added together in calls from some other part of the codebase. What we really want to know is "does it add", not "does it add 1+1".

In the present case, what we want to know is excatly 4 things:
1. Does it bail out early when we're not supposed to defer JS? We don't care why we can't, as that's for another method to determine and we might add/remove those conditions. So we're not testing the specifics. Just that it bails and returns empty array.

2. Does it include items from the default list? Again, we don't care how many or what the specific items are in the default list. Just that it includes them. So pick any item, and assert that it's included. Done.

3. Does it include items added from the user option? Again, we don't care what the user adds, only that it includes the item in the return array.

4. Does it de-duplicate items, should the user add an item that is already defered some other way?

The approach here is to construct 4 tests for those 4 concerns. Each has a cyclomatic complexity of 1 -- it will pass/fail for one reason only.

Note also that there are no fixtures necessary here. Each test tests a different thing. Fixtures are ideally for when you need sevaral data sets to cover instances of the same concern, and since we have 1 dataset for each concern, using a fixture and increasing the cyclomatic complexity of a single test to check for disparate concerns (as we have been doing) is a violation of separation of concerns.

As it is here, it's clear immediately from the test name in each case what is under test.

* Add missing specificity to shouldReceive('get')

* Fix test method names