Extract separate unit test for cannot specify case
Extracting a separate test for when image dimensions are not allowed to be specified (for whatever reason).

Why? The current test is too complex; trying to test too much. This makes it hard to read and combines testing different concerns.

What we want to know here is whether it bails out. The reasons for that bailout are specified by a different method and shouldn't be under test here.