Add test and update to force ratio to integer
This adds a testcase and solves for when getting the opposite dimension results in a float.
The HTML for width/height should specify a pixel value as an integer.

We use round() to round up or down based on the decimal value, then cast to integer specifically to get rid of the float type.