Revise bailout logic to make the new tests pass
Web stories creates a different `is_amp_endpoint()` whenever AMP's original function is not defined. When we get to this point in the code, we don't know whose version of the function we're evaluating.

So first, we can check that there is a function. If not, we can bail out -- nothing more to do.

Next, we have a function, but we don't know which. If whatever `is_amp_endpoint()` is false AND the web stories plugin not active, now we know that we can trust the `is_amp_endpoint()`'s telling us it's not -- web stories alternate criteria are not in play, so we can bail out.

Finally, we know that either it's a normal amp page, OR it's Web Stories is in play. So we check for web stories active AND this is a web-stories singular page.